run.c: Fix uninitialized field
The `fd` field in the `Transformer` structure is not initialized.
Because of that, on certain code paths, calls to the system functions
will use a wrong file descriptor. This causes an assertion failure in
Unikraft.

Workaround for this was the commenting out of the
`CONFIG_LIBUKDEBUG_ENABLE_ASSERT` configuration option in Unikraft.

Fix this by initializing `fd` to -1. This makes the current code fully
functional on Unikraft, eliminating the assertion failure.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>